secmet: rename feature module to features to match qualifiers
