+++ antismash/common/secmet/features/antismash_feature.py
