+++ antismash/common/secmet/features/cds_motif.py
