+++ antismash/common/secmet/features/feature.py
