+++ antismash/common/secmet/features/gene.py
