+++ antismash/modules/active_site_finder/analysis.py
@@ -15,13 +15,13 @@ from antismash.common import secmet
 from .common import ActiveSiteAnalysis
 
 # these are type aliases, not constants
-SinglePairing = Tuple[secmet.feature.Domain, str]  # pylint: disable=invalid-name
-MultiplePairing = Tuple[secmet.feature.Domain, List[str]]  # pylint: disable=invalid-name
+SinglePairing = Tuple[secmet.features.Domain, str]  # pylint: disable=invalid-name
+MultiplePairing = Tuple[secmet.features.Domain, List[str]]  # pylint: disable=invalid-name
 
 
 def run_all_analyses(record: secmet.Record) -> List[MultiplePairing]:
     """ Runs all AFS analyses at once and returns their aggregated results """
-    hits_by_feature = defaultdict(list)  # type: Dict[secmet.feature.Domain, List[str]]
+    hits_by_feature = defaultdict(list)  # type: Dict[secmet.features.Domain, List[str]]
     for analysis in [asp_ks, asp_ks_c, asp_at, acp_type, asp_acp, asp_pksi_dh, asp_pksi_kr,
                      asp_thioesterase, pksi_er_stereo, pksi_kr_stereo, pksi_at_spec, asp_p450_oxy]:
         for feature, hit in analysis(record):
